requirements.txt
absl-py==1.0.0
# argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1610522577486/work
# attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1640799537051/work
# backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
# backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
# beautifulsoup4 @ file:///home/conda/feedstock_root/build_artifacts/beautifulsoup4_1649463573192/work
# bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1649361991009/work
cachetools==5.0.0
certifi==2021.10.8
# cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1625835301720/work
charset-normalizer==2.0.12
colorlog==6.6.0
cupy-cuda101==9.6.0
cvxpy==1.1.18
cycler==0.11.0
debugpy==1.6.0
# decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
# defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
dlib==19.22.1
ecos==2.0.10
# entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
faiss-gpu==1.7.2
# fastjsonschema @ file:///home/conda/feedstock_root/build_artifacts/python-fastjsonschema_1641751198313/work/dist
fastrlock==0.8
fonttools==4.30.0
google-auth==2.6.5
google-auth-oauthlib==0.4.6
graphviz==0.8.4
grpcio==1.45.0
idna==3.3
imageio==2.16.1
# importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1648728291958/work
# importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1649682376682/work
# ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1649684273175/work/dist/ipykernel-6.13.0-py3-none-any.whl
# ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1645888813068/work
ipython-genutils==0.2.0
ipywidgets==7.7.0
# jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1649067102072/work
# Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1648299710939/work
joblib==1.1.0
# jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema-meta_1642000296051/work
jupyter-client==7.2.2
# jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1645024288521/work
# jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1649765532372/work
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
llvmlite==0.38.0
Markdown==3.3.6
# MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1621455667184/work
matplotlib==3.5.1
# matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
# mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1624941291602/work
mpmath==1.2.1
mxnet-cu101==1.7.0
# nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1649785910383/work
# nbconvert @ file:///home/conda/feedstock_root/build_artifacts/nbconvert-meta_1649676641343/work
# nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1649105480291/work
# nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1648959695634/work
networkx==2.6.3
# notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1647377876077/work
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
osqp==0.6.2.post5
# packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.3.5
# pandocfilters @ file:///home/conda/feedstock_root/build_artifacts/pandocfilters_1631603243851/work
# parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
# pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
# pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.1.0
# prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1649447152425/work
# prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1649130487073/work
protobuf==3.20.0
psutil==5.9.0
# ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
# pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
# Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
# pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1649603503565/work
# pyrsistent @ file:///home/conda/feedstock_root/build_artifacts/pyrsistent_1610146801408/work
# python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2022.1
PyWavelets==1.3.0
pyzmq==22.3.0
qdldl==0.1.5.post0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
scs==3.2.0
seaborn==0.11.2
# Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
# six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
# soupsieve @ file:///home/conda/feedstock_root/build_artifacts/soupsieve_1638550740809/work
sympy==1.10.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
# terminado @ file:///home/conda/feedstock_root/build_artifacts/terminado_1649103772573/work
threadpoolctl==3.1.0
tifffile==2021.11.2
# tinycss2 @ file:///home/conda/feedstock_root/build_artifacts/tinycss2_1637612658783/work
torch==1.8.0+cu101
torch-tb-profiler==0.4.0
torchaudio==0.8.0
torchvision==0.9.0+cu101
# tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094704591/work
tqdm==4.64.0
# traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
tsnecuda==2.1.0
# typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
urllib3==1.26.9
# wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
# zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1649012893348/work
